refactor(#741): refactor identifierTypeController
