--- api/controllers/v1/IdentifierTypeController.js
@@ -1,8 +1,0 @@
-/**
- */
-
-const identifierTypeController = require('../IdentifierTypeController');
-
-module.exports = {
-  findAll: (req, res) => identifierTypeController.findAll(req, res),
-};
