+++ api/controllers/v1/file-format/find-all.js
@@ -1,0 +1,10 @@
+module.exports = async (req, res) => {
+  try {
+    const fileFormats = await TFileFormat.find();
+    return res.ok(fileFormats);
+  } catch (err) {
+    return res.serverError(
+      'There was a problem while retrieving the file formats.'
+    );
+  }
+};
